use gorm.io/driver/sqlite instead of cloudquery/sqlite to fix #12

+++ db/map.go
@@ -2,10 +2,10 @@ package db
 
 import "github.com/ssst0n3/awesome_libs"
 
-func MapObjectById(objects []map[string]interface{}) map[int64]awesome_libs.Dict {
-	result := map[int64]awesome_libs.Dict{}
+func MapObjectById(objects []map[string]interface{}) map[uint]awesome_libs.Dict {
+	result := map[uint]awesome_libs.Dict{}
 	for _, object := range objects {
-		id := object["id"].(int64)
+		id := object["id"].(uint)
 		result[id] = object
 	}
 	return result
